refactor(ai): migrate list_discussions to tools registry
