+++ src/lib/ai/tools/executor.ts
@@ -160,12 +160,6 @@ const NON_ADMIN_RLS_READ_TOOL_NAMES: ReadonlySet<ToolName> = new Set<ToolName>([
   "find_navigation_targets",
 ]);
 
-const listDiscussionsSchema = z
-  .object({
-    limit: z.number().int().min(1).max(25).optional(),
-  })
-  .strict();
-
 const listJobPostingsSchema = z
   .object({
     limit: z.number().int().min(1).max(25).optional(),
@@ -395,7 +389,6 @@ const findNavigationTargetsSchema = z
   .strict();
 
 const ARG_SCHEMAS: Partial<Record<ToolName, z.ZodSchema>> = {
-  list_discussions: listDiscussionsSchema,
   list_job_postings: listJobPostingsSchema,
   list_chat_groups: listChatGroupsSchema,
   list_alumni: listAlumniSchema,
@@ -647,40 +640,6 @@ interface EventValidationErrorRecord {
   index: number;
   missing_fields: string[];
   draft: Record<string, unknown>;
-}
-
-async function listDiscussions(
-  sb: SB,
-  orgId: string,
-  args: z.infer<typeof listDiscussionsSchema>,
-  logContext: AiLogContext
-): Promise<ToolExecutionResult> {
-  const limit = Math.min(args.limit ?? 10, 25);
-  return safeToolQuery(logContext, async () => {
-    const { data, error } = await sb
-      .from("discussion_threads")
-      .select("id, title, body, author_id, reply_count, created_at")
-      .eq("organization_id", orgId)
-      .is("deleted_at", null)
-      .order("created_at", { ascending: false })
-      .limit(limit);
-
-    if (!Array.isArray(data) || error) {
-      return { data, error };
-    }
-
-    return {
-      data: data.map((discussion) => ({
-        id: discussion.id,
-        title: discussion.title,
-        author_id: discussion.author_id,
-        reply_count: discussion.reply_count ?? 0,
-        created_at: discussion.created_at ?? null,
-        body_preview: truncateBody(discussion.body),
-      })),
-      error: null,
-    };
-  });
 }
 
 async function listJobPostings(
@@ -3083,13 +3042,6 @@ export async function executeToolCall(
         return dispatchToolModule(toolName, args, { ctx, sb, logContext });
       }
       switch (toolName) {
-        case "list_discussions":
-          return listDiscussions(
-            sb,
-            ctx.orgId,
-            args as z.infer<typeof listDiscussionsSchema>,
-            logContext
-          );
         case "list_job_postings":
           return listJobPostings(
             sb,
